Combined rammon and cpumon to sysmon, and upgraded them
