+++ snakewm/apps/system/cpumon/ramproc.py
@@ -25,7 +25,7 @@ def ramproc2():
 
     total = isolate_num(file[0])
     used = total - isolate_num(file[2])  # uses memavaiable line
-    return "{}/{}".format(used, total)
+    return "Ram Usage: <br>{}/{}".format(used, total)
 
 
 # Test the module
